move NodeLock into separate file
